# More optimization for x86 AVX