--- a/dlk/python/dlk/templates/src/func/impl/x86_avx/apply_thresholds.cpp
+++ b/dlk/python/dlk/templates/src/func/impl/x86_avx/apply_thresholds.cpp
@@ -24,15 +24,18 @@ namespace dlk {
 
 namespace impl {
 
+constexpr std::size_t MAX_CHANNELS = 4096;
+
+static const auto buf_ts0 = std::make_unique<BIN_CONV_OUTPUT[]>(MAX_CHANNELS);
+static const auto buf_ts1 = std::make_unique<BIN_CONV_OUTPUT[]>(MAX_CHANNELS);
+static const auto buf_ts2 = std::make_unique<BIN_CONV_OUTPUT[]>(MAX_CHANNELS);
+static const auto buf_flg = std::make_unique<BIN_CONV_OUTPUT[]>(MAX_CHANNELS);
+
+
 void ApplyThresholds(
     dlk::MatrixView<BIN_CONV_OUTPUT, dlk::MatrixOrder::ColMajor> &result,
     const binary_convolution_parameters &p) {
   Measurement::Start("ApplyThresholds");
-
-  const auto buf_ts0 = std::make_unique<BIN_CONV_OUTPUT[]>(result.rows());
-  const auto buf_ts1 = std::make_unique<BIN_CONV_OUTPUT[]>(result.rows());
-  const auto buf_ts2 = std::make_unique<BIN_CONV_OUTPUT[]>(result.rows());
-  const auto buf_flg = std::make_unique<BIN_CONV_OUTPUT[]>(result.rows());
 
   for (unsigned int i = 0; i < result.rows(); ++i) {
     T_INT ts0 = p.thresholds[NUM_OF_A2W1_THRESHOLD * i];
@@ -79,11 +82,6 @@ void ApplyThresholdsAndPack(
     QUANTIZED_PACKED output[]) {
   Measurement::Start("ApplyThresholdsAndPack");
 
-  const auto buf_ts0 = std::make_unique<BIN_CONV_OUTPUT[]>(result.rows());
-  const auto buf_ts1 = std::make_unique<BIN_CONV_OUTPUT[]>(result.rows());
-  const auto buf_ts2 = std::make_unique<BIN_CONV_OUTPUT[]>(result.rows());
-  const auto buf_flg = std::make_unique<BIN_CONV_OUTPUT[]>(result.rows());
-
   for (unsigned int i = 0; i < result.rows(); ++i) {
     T_INT ts0 = p.thresholds[NUM_OF_A2W1_THRESHOLD * i];
     T_INT ts1 = p.thresholds[NUM_OF_A2W1_THRESHOLD * i + 1];
@@ -100,8 +98,65 @@ void ApplyThresholdsAndPack(
     buf_flg[i] = flag;
   }
 
+  const auto count = result.cols();
+  const auto count_floor = count - (count % 4);
+
 #pragma omp parallel for
-  for (unsigned int j = 0; j < result.cols(); ++j) {
+  for (unsigned int j = 0; j < count_floor; j += 4) {
+    for (unsigned int i = 0; i < result.rows(); i += 32) {
+#define LOAD_TH(k) \
+      const auto ts0##k = _mm256_loadu_si256(reinterpret_cast<__m256i*>(buf_ts0.get() + i + k * 16)); \
+      const auto ts1##k = _mm256_loadu_si256(reinterpret_cast<__m256i*>(buf_ts1.get() + i + k * 16)); \
+      const auto ts2##k = _mm256_loadu_si256(reinterpret_cast<__m256i*>(buf_ts2.get() + i + k * 16)); \
+      const auto flg##k = _mm256_loadu_si256(reinterpret_cast<__m256i*>(buf_flg.get() + i + k * 16)); \
+      const auto is_neg##k = _mm256_cmpgt_epi16(_mm256_setzero_si256(), flg##k); \
+      const auto m2##k = _mm256_sub_epi16(flg##k, _mm256_set1_epi16(2)); \
+      const auto is_not_const##k = _mm256_cmpgt_epi16(_mm256_setzero_si256(), m2##k);
+
+#define APPLY(k, l) \
+      const auto d##k##l = _mm256_loadu_si256(reinterpret_cast<__m256i*>(result.data(i + k * 16, j + l))); \
+      const auto f0##k##l = _mm256_andnot_si256(_mm256_cmpgt_epi16(ts0##k, d##k##l), flg##k); \
+      const auto f1##k##l = _mm256_andnot_si256(_mm256_cmpgt_epi16(ts1##k, d##k##l), flg##k); \
+      const auto f2##k##l = _mm256_andnot_si256(_mm256_cmpgt_epi16(ts2##k, d##k##l), flg##k); \
+      const auto tmp##k##l = _mm256_add_epi16(_mm256_add_epi16(f0##k##l, f1##k##l), _mm256_add_epi16(f2##k##l, is_neg##k)); \
+      const auto res##k##l = _mm256_blendv_epi8(m2##k, tmp##k##l, is_not_const##k);
+
+#define PACK(l) \
+      const auto packed##l = _mm256_packs_epi16(res0##l, res1##l); \
+      const auto permuted##l = _mm256_permute4x64_epi64(packed##l, 0xD8); \
+      const auto vlsb##l = _mm256_slli_epi32(permuted##l, 7); \
+      const auto vmsb##l = _mm256_slli_epi32(permuted##l, 6); \
+      const auto lsb##l = _mm256_movemask_epi8(vlsb##l); \
+      const auto msb##l = _mm256_movemask_epi8(vmsb##l);
+
+      LOAD_TH(0)
+      LOAD_TH(1)
+      APPLY(0, 0)
+      APPLY(1, 0)
+      APPLY(0, 1)
+      APPLY(1, 1)
+      APPLY(0, 2)
+      APPLY(1, 2)
+      APPLY(0, 3)
+      APPLY(1, 3)
+      PACK(0)
+      PACK(1)
+      PACK(2)
+      PACK(3)
+      const auto index = (j + (i / 32) * result.cols()) * 2;
+      output[index+0] = QUANTIZED_PACKED(lsb0);
+      output[index+1] = QUANTIZED_PACKED(msb0);
+      output[index+2] = QUANTIZED_PACKED(lsb1);
+      output[index+3] = QUANTIZED_PACKED(msb1);
+      output[index+4] = QUANTIZED_PACKED(lsb2);
+      output[index+5] = QUANTIZED_PACKED(msb2);
+      output[index+6] = QUANTIZED_PACKED(lsb3);
+      output[index+7] = QUANTIZED_PACKED(msb3);
+#undef APPLY
+    }
+  }
+
+  for (unsigned int j = count_floor; j < count; ++j) {
     for (unsigned int i = 0; i < result.rows(); i += 32) {
 #define APPLY(k) \
       const auto d##k = _mm256_loadu_si256(reinterpret_cast<__m256i*>(result.data(i + k * 16, j))); \