… dlk/python/dlk/templates/src/func/impl/x86_avx/quantized_conv2d_tiling.cpp