<!--- 👉👉👉 Provide a general summary of your changes. 👈👈👈 -->
| |Core i7-8750H@governor=performance|Xeon Gold 6154|
|----|----|----|
|before|49.8ms|100ms|
|by #365 |2.34ms@OMP_NUM_THREADS=6|7.43ms@OMP_NUM_THREADS=2|
|This PR|1.73ms@OMP_NUM_THREADS=6|1.85ms@OMP_NUM_THREADS=18|

Execution time varies, these are minimum time.

## Motivation and Context
<!--- 💭💡💭 Why is this change required? What problem does it solve? 💭💡💭 -->
<!--- 🔗 If it fixes an open issue, please link to the issue here. 🔗 -->
This PR provides more optimization for x86 AVX environments.

## Description
<!--- 📝🎯📝 Describe your changes in detail. 📝🎯📝 -->
- Optimize lookup
- Optimize extract_image_patches
- Optimize quantized_conv2d_tiling
- Initialize OpenMP thread pool in Network::init()
- Optimize ApplyThresholdsAndPack

## How has this been tested?
<!--- 🙆👌🙆 Please describe in detail how you tested your changes. 🙆👌🙆 -->
<!--- Include details of your testing environment, details of your manual tests, snippet code or commands of your manual tests, table of experiments results metrics, tests ran to see how your change affects other areas of the code, etc. -->
<!--- If you have experiments report documents please link to here. -->
```shell
$ PYTHONPATH=python/dlk python python/dlk/scripts/generate_project.py -i minimal_graph_with_shape.pb -o tmp/ -p object_detection -hq -ts
$ cd tmp/object_detection.prj
$ CXXFLAGS=-DFUNC_TIME_MEASUREMENT make -j8 lm_x86_avx
```

## Screenshots (if appropriate):

## Types of changes
<!--- 🏁🎌🏁 What types of changes does your code introduce? Put an `x` in all the boxes that apply: 🏁🎌🏁 -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature / Optimization (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

## Checklist:
<!--- ✅✅✅ Go over all the following points, and put an `x` in all the boxes that apply. ✅✅✅ -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

<!--- TODO(wakisaka): After decide our code style, add this.
- [ ] My code follows the code style of this project.
-->
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.