@iizukak 2019-07-30
@primenumber 
```
$ ./lm_x86.elf network_input_output/000_images_placeholder:0.npy network_input_output/337_output:0.npy
-------------------------------------------------------------
Comparison: Default network test  succeeded!!!
-------------------------------------------------------------
TotalInitTime 27970,  sum:27.97ms
TotalRunTime 144468,  sum:144.468ms
..Convolution 8940,218,  sum:9.158ms
....kn2row-1x1 8933,217,  sum:9.15ms
......matrix_multiplication 8930,216,  sum:9.146ms
..BatchNorm 4383,18,  sum:4.401ms
..QTZ_linear_mid_tread_half 11721,  sum:11.721ms
....pack_input 6234,  sum:6.234ms
..QuantizedConv2D 65044,17687,8308,3876,4306,2541,1466,223,302,786,6422,1727,1208,2560,2462,  sum:118.918ms
....Tensor convert 46,76,35,25,9,5,1,4,9,17,8,2,2,2,1,  sum:0.242ms
....quantized-kn2row 64993,17569,8270,3848,4294,2534,1463,217,291,768,6413,1723,1205,2556,2459,  sum:118.603ms
......quantized_matrix_multiplication 37919,12735,6951,3352,3989,2436,1291,187,204,446,6250,1605,1036,2448,2331,  sum:83.18ms
......matrix_shift_add1 819,312,212,124,133,51,103,95,64,95,49,95,  sum:2.152ms
......matrix_shift_add2 5118,1128,415,108,43,8,15,14,11,13,8,14,  sum:6.895ms
......ApplyThresholds 3748,1031,414,144,69,16,26,18,52,197,27,24,32,25,  sum:5.823ms
......pack bits 2163,491,175,58,35,7,14,7,28,112,13,8,13,7,  sum:3.131ms
..Memcpy 42,5,5,1,1,0,0,0,1,2,0,0,0,0,  sum:0.057ms
..ExtractImagePatches 61,17,4,4,2,1,1,4,  sum:0.094ms
..func_ConcatOnDepth 10,  sum:0.01ms
..QuantizedConv2D_ApplyScalingFactor 14,  sum:0.014ms
..Add 6,1,  sum:0.007ms
..Max 29,  sum:0.029ms
```
```
$ OMP_NUM_THREADS=4 ./lm_x86_avx.elf network_input_output/000_images_placeholder:0.npy network_input_output/337_output:0.npy
-------------------------------------------------------------
Comparison: Default network test  succeeded!!!
-------------------------------------------------------------
TotalInitTime 30818,  sum:30.818ms
TotalRunTime 32055,  sum:32.055ms
..Convolution 11418,344,  sum:11.762ms
....kn2row-1x1 11411,343,  sum:11.754ms
......matrix_multiplication 11407,341,  sum:11.748ms
..BatchNorm 4216,30,  sum:4.246ms
..QTZ_linear_mid_tread_half 10607,  sum:10.607ms
....pack_input 6237,  sum:6.237ms
..QuantizedConv2D 2263,459,338,468,220,163,90,21,27,46,647,87,89,86,100,  sum:5.104ms
....Tensor convert 409,6,4,1,1,0,0,0,1,2,0,0,0,0,0,  sum:0.424ms
....Quantized Conv2D Tiling 1672,407,316,452,207,154,81,14,15,28,636,79,80,79,80,  sum:4.3ms
....ApplyThresholdsAndPack 143,31,11,10,7,5,6,4,6,11,7,5,6,5,  sum:0.257ms
....copy 31,9,3,1,1,0,0,0,1,2,0,0,0,0,3,  sum:0.051ms
....Output tensor convert 14,  sum:0.014ms
..Memcpy 45,7,3,1,1,0,0,0,1,2,0,0,0,0,  sum:0.06ms
..ExtractImagePatches 73,22,3,3,1,0,1,7,  sum:0.11ms
..func_ConcatOnDepth 15,  sum:0.015ms
..QuantizedConv2D_ApplyScalingFactor 19,  sum:0.019ms
..Add 22,3,  sum:0.025ms
..Max 47,  sum:0.047ms
```

So nice...