@hadusam 2019-07-30
run blueoil test